Added a very basic UI

+++ DBMSExplorer/Packages/DBMSKit/Sources/DBMSKit/Helpers/StringHelpers.swift
@@ -127,12 +127,12 @@ class StringHelpers {
         return newString
     }
     
-    static func sdbmHash(str: String) -> UInt64 {
-        var hash: UInt64 = 0
+    static func sdbmHash(str: String) -> UInt16 {
+        var hash: UInt16 = 0
         let characters = Array(str)
 
         for i in 0..<str.count {
-            hash = UInt64(characters[i].asciiValue ?? 0) &+ (hash << 6) &+ (hash << 16) &- hash
+            hash = UInt16(characters[i].asciiValue ?? 0) &+ (hash << 6) &+ (hash << 16) &- hash
         }
 
         return hash
